Add network library (developing...)

+++ scripts/core_common/modules/glew.py
@@ -18,14 +18,17 @@ def make():
   if not config.check_option("module", "mobile"):
     return;
 
+  print("[fetch & build]: glew")
   base_dir = base.get_script_dir() + "/../../core/Common/3dParty/glew"
   old_cur = os.getcwd()
   os.chdir(base_dir)
 
   base.common_check_version("glew", "1", clean)
-  base.download("https://deac-ams.dl.sourceforge.net/project/glew/glew/2.1.0/glew-2.1.0-win32.zip", "./archive.zip")
-  base.extract("./archive.zip", "./")
-  base.delete_file("./archive.zip")
+
+  if not base.is_dir("glew-2.1.0"):
+    base.download("https://deac-ams.dl.sourceforge.net/project/glew/glew/2.1.0/glew-2.1.0-win32.zip", "./archive.zip")
+    base.extract("./archive.zip", "./")
+    base.delete_file("./archive.zip")
 
   os.chdir(old_cur)
   return
